requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
Ollama
sentence_transformers
tokenizers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
langchain_groq
langchain_core
fastapi
langserve
sse_starlette
pydantic==1.10.13
sentence-transformers
pandas
dill==0.3.7